+++ python/lib/sift_py/rule/config.py
@@ -4,11 +4,11 @@ from abc import ABC, abstractmethod
 from enum import Enum
 from typing import Any, Dict, List, Optional, TypedDict, Union, cast
 
+from sift.annotations.v1.annotations_pb2 import AnnotationType
 from sift.rules.v1.rules_pb2 import ActionKind
 
 from sift_py._internal.convert.json import AsJson
 from sift_py.ingestion.channel import ChannelConfig
-from sift_py.yaml.rule import RuleActionAnnotationKind
 
 
 class RuleConfig(AsJson):
@@ -185,6 +185,26 @@ class RuleActionKind(Enum):
             return cls.ANNOTATION
 
         return None
+
+
+class RuleActionAnnotationKind(Enum):
+    REVIEW = "review"
+    PHASE = "phase"
+
+    @classmethod
+    def from_annotation_type(cls, annotation_type: AnnotationType) -> "RuleActionAnnotationKind":
+        if annotation_type == AnnotationType.ANNOTATION_TYPE_PHASE:
+            return cls.PHASE
+        return cls.PHASE
+
+    @classmethod
+    def from_str(cls, val: str) -> "RuleActionAnnotationKind":
+        if val == cls.REVIEW.value:
+            return cls.REVIEW
+        elif val == cls.PHASE.value:
+            return cls.PHASE
+        else:
+            raise ValueError("Argument 'val' is not a valid annotation kind.")
 
 
 class RuleActionKindStrRep(Enum):
