+++ python/lib/sift_py/yaml/rule.py
@@ -1,16 +1,15 @@
 from __future__ import annotations
 
 import re
-from enum import Enum
+from dataclasses import dataclass
 from pathlib import Path
 from typing import Any, Dict, List, Literal, Union, cast
 
 import yaml
-from sift.annotations.v1.annotations_pb2 import AnnotationType
 from typing_extensions import NotRequired, TypedDict
 
 from sift_py.ingestion.config.yaml.error import YamlConfigError
-from sift_py.rule.service import SubExpression
+from sift_py.rule.config import RuleActionAnnotationKind
 from sift_py.yaml.channel import (
     ChannelConfigYamlSpec,
     _validate_channel_reference,
@@ -342,21 +341,11 @@ class NamedExpressionYamlSpec(TypedDict):
     name: str
 
 
-class RuleActionAnnotationKind(Enum):
-    REVIEW = "review"
-    PHASE = "phase"
+@dataclass
+class SubExpression:
+    fully_qualified_rule_name: str
+    expressions: Dict[str, Any]
 
-    @classmethod
-    def from_annotation_type(cls, annotation_type: AnnotationType) -> "RuleActionAnnotationKind":
-        if annotation_type == AnnotationType.ANNOTATION_TYPE_PHASE:
-            return cls.PHASE
-        return cls.PHASE
-
-    @classmethod
-    def from_str(cls, val: str) -> "RuleActionAnnotationKind":
-        if val == cls.REVIEW.value:
-            return cls.REVIEW
-        elif val == cls.PHASE.value:
-            return cls.PHASE
-        else:
-            raise ValueError("Argument 'val' is not a valid annotation kind.")
+    def __init__(self, fully_qualified_rule_name: str, expressions: Dict[str, Any]):
+        self.fully_qualified_rule_name = fully_qualified_rule_name
+        self.expressions = expressions
